fix Script

+++ Ansite/installer.sh
@@ -25,27 +25,27 @@ if which opkg > /dev/null 2>&1; then
 fi
 
 if python --version 2>&1 | grep -q '^Python 3\.'; then
-    echo "You have Python3 image"
+    echo ":You have Python3 image ..."
     sleep 1; clear
     VERSION='1.4'
 else
-    echo "You have Python2 image"
+    echo ":You have Python2 image ..."
     sleep 1; clear
     VERSION='1.3'
 fi
 
 ##################################
 # Remove previous files (if any) #
-rm -rf $TMPDIR/${PACKAGE}*
+rm -rf $TMPDIR/"${PACKAGE:?}/"* > /dev/null 2>&1
 
 ######################
 #  Remove Old Plugin #
 if grep -qs "Package: $PACKAGE" $STATUS ; then
-    echo "Remove old version..."
+    echo "   >>>>   Remove old version   <<<<"
     $OPKGREMOV $PACKAGE
     sleep 1; clear
 else
-    echo "No older version was found on the device... "
+    echo "   >>>>   No Older Version Was Found   <<<<"
     sleep 1; clear
 fi
 
@@ -55,18 +55,18 @@ $OPKG > /dev/null 2>&1
 ###################
 #  Install Plugin #
 if python --version 2>&1 | grep -q '^Python 3\.'; then
-    echo "Downloading And Insallling Ansite plugin Please Wait ......"
+    echo "Insallling Ansite plugin Please Wait ......"
     wget $MY_URL/${PACKAGE}_${VERSION}_all.ipk -qP $TMPDIR
     $OPKGINSTAL $TMPDIR/${PACKAGE}_${VERSION}_all.ipk
 else
-    echo "Downloading And Insallling Ansite plugin Please Wait ......"
+    echo "Insallling Ansite plugin Please Wait ......"
     wget $MY_URL/${PACKAGE}_${VERSION}_all.ipk -qP $TMPDIR
     $OPKGINSTAL $TMPDIR/${PACKAGE}_${VERSION}_all.ipk
 fi
 
 #########################
 # Remove files (if any) #
-rm -rf $TMPDIR/${PACKAGE}*
+rm -rf $TMPDIR/"${PACKAGE:?}/"* > /dev/null 2>&1
 
 
 sleep 2; clear
