+++ Suptv/installer.sh
@@ -23,43 +23,43 @@ if which opkg > /dev/null 2>&1; then
 fi
 
 if python --version 2>&1 | grep -q '^Python 3\.'; then
-    echo "You have Python3 image"
+    echo ":You have Python3 image ..."
     sleep 1; clear
     VERSION='3.1'
 else
-    echo "You have Python2 image"
+    echo ":You have Python2 image ..."
     sleep 1; clear
     VERSION='2.0.1'
 fi
 
 ##################################
 # Remove previous files (if any) #
-rm -rf $TMPDIR/${PACKAGE}*
+rm -rf $TMPDIR/"${PACKAGE:?}/"* > /dev/null 2>&1
 
 ###################
 #  Install Plugin #
 
 if grep -qs "Package: $PACKAGE" $STATUS ; then
-    echo "SupTV found in device..."
+    echo "   >>>>   SupTV found in device   <<<<"
     sleep 1; clear
     echo "Goodbye!"
 elif python --version 2>&1 | grep -q '^Python 3\.'; then
     echo "Opkg Update ..."
     $OPKG > /dev/null 2>&1
-    echo "Downloading And Insallling Suptv plugin Please Wait ......"
+    echo "Insallling Suptv plugin Please Wait ......"
     wget $MY_URL/${PACKAGE}_${VERSION}_all.ipk -qP $TMPDIR
     $OPKGINSTAL $TMPDIR/${PACKAGE}_${VERSION}_all.ipk
 else
     echo "Opkg Update ..."
     $OPKG > /dev/null 2>&1
-    echo "Downloading And Insallling Suptv plugin Please Wait ......"
+    echo "Insallling Suptv plugin Please Wait ......"
     wget $MY_URL/${PACKAGE}_${VERSION}_all.ipk -qP $TMPDIR
     $OPKGINSTAL $TMPDIR/${PACKAGE}_${VERSION}_all.ipk
 fi
 
 ##################################
 # Remove previous files (if any) #
-rm -rf $TMPDIR/${PACKAGE}*
+rm -rf $TMPDIR/"${PACKAGE:?}/"* > /dev/null 2>&1
 
 sleep 2; clear
 echo ""
